fix(added a mock to the tests)
The tests were trying to check if a snippet was existing, but it wasn't so it returned an error to the test. The same happened when it checked if the snippet didn't exist, because it actually existed.

In order to fix that, I've added two mocks to the tests so we do not modify the production storage and the tests do actually pass.