Fix the unit tests
Fix the unit tests from issue #5.

The first test case is looking to demonstrate functionality for a
snippet that exists but the chosen snippet doesn't exist in storage.

The second test case is checking if we get the correct error code when
a snippet does not exist. unfortunately the chosen snippet does exist in
storage.

Both problems can be fixed by using a mock.